1.04.24

+++ src/main/java/ru/BouH/engine/game/resources/assets/models/basic/forms/D2/VectorModel2D.java
@@ -1,16 +1,16 @@
 package ru.BouH.engine.game.resources.assets.models.basic.forms.D2;
 
-import org.joml.Vector2d;
+import org.joml.Vector2f;
 import ru.BouH.engine.game.resources.assets.models.Model;
 import ru.BouH.engine.game.resources.assets.models.basic.forms.BasicMesh;
 import ru.BouH.engine.game.resources.assets.models.formats.Format2D;
 import ru.BouH.engine.game.resources.assets.models.mesh.Mesh;
 
 public class VectorModel2D implements BasicMesh<Format2D> {
-    private final Vector2d v1;
-    private final Vector2d v2;
+    private final Vector2f v1;
+    private final Vector2f v2;
 
-    public VectorModel2D(Vector2d v1, Vector2d v2) {
+    public VectorModel2D(Vector2f v1, Vector2f v2) {
         this.v1 = v1;
         this.v2 = v2;
     }
@@ -23,11 +23,11 @@ public class VectorModel2D implements BasicMesh<Format2D> {
     @Override
     public Mesh generateMesh() {
         Mesh mesh = new Mesh();
-        mesh.putPositionValue((float) this.v1.x);
-        mesh.putPositionValue((float) this.v1.y);
+        mesh.putPositionValue(this.v1.x);
+        mesh.putPositionValue(this.v1.y);
 
-        mesh.putPositionValue((float) this.v2.x);
-        mesh.putPositionValue((float) this.v2.y);
+        mesh.putPositionValue(this.v2.x);
+        mesh.putPositionValue(this.v2.y);
 
         mesh.putIndexValue(0);
         mesh.putIndexValue(1);
