+++ src/main/java/ru/BouH/engine/game/resources/assets/models/basic/forms/D3/VectorModel3D.java
@@ -1,6 +1,6 @@
 package ru.BouH.engine.game.resources.assets.models.basic.forms.D3;
 
-import org.joml.Vector3d;
+import org.joml.Vector3f;
 import ru.BouH.engine.game.resources.assets.materials.Material;
 import ru.BouH.engine.game.resources.assets.models.Model;
 import ru.BouH.engine.game.resources.assets.models.basic.forms.BasicMesh;
@@ -8,10 +8,10 @@ import ru.BouH.engine.game.resources.assets.models.formats.Format3D;
 import ru.BouH.engine.game.resources.assets.models.mesh.Mesh;
 
 public class VectorModel3D implements BasicMesh<Format3D> {
-    private final Vector3d v1;
-    private final Vector3d v2;
+    private final Vector3f v1;
+    private final Vector3f v2;
 
-    public VectorModel3D(Vector3d v1, Vector3d v2) {
+    public VectorModel3D(Vector3f v1, Vector3f v2) {
         this.v1 = v1;
         this.v2 = v2;
     }
@@ -24,13 +24,13 @@ public class VectorModel3D implements BasicMesh<Format3D> {
     @Override
     public Mesh generateMesh() {
         Mesh mesh = new Mesh();
-        mesh.putPositionValue((float) this.v1.x);
-        mesh.putPositionValue((float) this.v1.y);
-        mesh.putPositionValue((float) this.v1.z);
+        mesh.putPositionValue(this.v1.x);
+        mesh.putPositionValue(this.v1.y);
+        mesh.putPositionValue(this.v1.z);
 
-        mesh.putPositionValue((float) this.v2.x);
-        mesh.putPositionValue((float) this.v2.y);
-        mesh.putPositionValue((float) this.v2.z);
+        mesh.putPositionValue(this.v2.x);
+        mesh.putPositionValue(this.v2.y);
+        mesh.putPositionValue(this.v2.z);
 
         mesh.putIndexValue(0);
         mesh.putIndexValue(1);
